requirements.txt
# pytorch / deep learning
torch
torchvision
pytorch-lightning
torchmetrics
geomloss
wandb
onnx
onnxruntime

# Specific for the neutrinos
nflows
pyunfold

# hydra
hydra-core
hydra-colorlog
hydra-optuna-sweeper

# plotting
matplotlib
seaborn
plotly

# data science
numpy
scipy
scikit-learn
scikit-hep
pandas
tables
h5py
pykeops

# other
black
pylint
dotmap
pyrootutils
pre-commit
rich
